+++ cypress/integration/core/autosuggest.spec.ts
@@ -315,14 +315,16 @@ describe("Autosuggest", () => {
   });
 
   it('should delay progress indicator when loadingDelayed is set', () => {
-    cy.get("dso-autosuggest").invoke("attr", "loading", true);
-    cy.get("dso-autosuggest").invoke("attr", "loading-delayed", 1000);
-    cy.get("@input").focus().type("ams");
-    cy.wait(200);
-    cy.get("@listbox").should("not.exist");
-    cy.get("dso-progress-indicator").should("not.be.visible");
-    cy.wait(1000);
-    cy.get("dso-progress-indicator").should("be.visible");
+    cy.visit('http://localhost:56106/iframe.html?id=autosuggest--loading-delayed');
+    cy.get('@input').focus().type('ams')
+      .wait(200)
+      .get('dso-autosuggest').invoke('attr', 'loading', true)
+      .get('dso-progress-indicator').should('not.exist')
+      .wait(1000)
+      .get('dso-progress-indicator').should('be.visible')
+      .wait(1500)
+      .get('dso-autosuggest').invoke('attr', 'loading', false)
+      .get('dso-progress-indicator').should('not.exist');
   });
 
   it('should show not found text when no results are found', () => {