+++ packages/sources/src/components/autosuggest/autosuggest.stories.ts
@@ -20,6 +20,7 @@ type AutosuggestTemplateFnType<TemplateFnReturnType> = (
 
 export interface AutosuggestParameters<TemplateFnReturnType> {
   autosuggestDemoTemplate: AutosuggestTemplateFnType<TemplateFnReturnType>;
+  autosuggestLoadingDelayedTemplate: AutosuggestTemplateFnType<TemplateFnReturnType>;
   autosuggestInSearchBarTemplate?: AutosuggestTemplateFnType<TemplateFnReturnType>;
 }
 
@@ -31,6 +32,7 @@ export function storiesOfAutosuggest<TemplateFnReturnType>(
   }: StorybookParameters,
   {
     autosuggestDemoTemplate,
+    autosuggestLoadingDelayedTemplate,
     autosuggestInSearchBarTemplate
   }: AutosuggestParameters<TemplateFnReturnType>
 ) {
@@ -54,6 +56,23 @@ export function storiesOfAutosuggest<TemplateFnReturnType>(
       return autosuggestDemoTemplate(fetchSuggestions, args.onSelect, args.onChange, args.onSearch, args.suggestOnFocus, args.loading, args.loadingLabel, args.loadingDelayed, args.notFoundLabel);
     }
   );
+
+  if (autosuggestLoadingDelayedTemplate) {
+    stories.add(
+      'loading delayed',
+      (a: Args | undefined) => {
+        if (!a) {
+          throw new ArgsError();
+        }
+
+        const args = a as AutosuggestArgs;
+
+        args.loadingDelayed = 1000;
+
+        return autosuggestLoadingDelayedTemplate(fetchSuggestions, args.onSelect, args.onChange, args.onSearch, args.suggestOnFocus, args.loading, args.loadingLabel, args.loadingDelayed, args.notFoundLabel);
+      }
+    );
+  }
 
   if (autosuggestInSearchBarTemplate) {
     stories.add(